pyproject.toml
[project]
name = "text-to-sql-generator"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.12"
dependencies = [
    "anthropic>=0.85.0",
    "python-dotenv>=1.2.2",
]

[dependency-groups]
dev = [
    "pytest>=9.0.2",
]
